+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - `f-input` 和 `f-textarea` 组件新增 `font-size` 配置项
 - `f-dialog` 组件新增 `height` 配置项
+- `f-popup` 组件新增 `fullscreen` 配置项
 
 **修复**